last commit

+++ src/BackEnd/Services/ProductService.cs
@@ -1,6 +1,10 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using TsuShopWebApi.Entities;
 using TsuShopWebApi.Interfaces;
+using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
+using Microsoft.Extensions.Options;
+using TsuShopWebApi.Options;
 
 namespace TsuShopWebApi.Services;
 
@@ -9,11 +13,17 @@ public class ProductService : IProductService
     
     private readonly IApplicationDbContext _context;
     private readonly IUserService _userService;
+    private readonly BlobServiceClient _blobServiceClient;
+    private readonly string _blobContainerName;
 
-    public ProductService(IApplicationDbContext context, IUserService userService)
+    public ProductService(IApplicationDbContext context,
+        IUserService userService,
+        IOptions<AzureBlobStorageOptions> azureBlobServiceOptions)
     {
         _context = context;
         _userService = userService;
+        _blobServiceClient = new BlobServiceClient(azureBlobServiceOptions.Value.ConnectionString);
+        _blobContainerName = azureBlobServiceOptions.Value.ContainerName;
     }
     
     
@@ -22,7 +32,6 @@ public class ProductService : IProductService
 
         var productsQuery = _context.Products.AsNoTracking().AsQueryable().Where(product => product.Price <= maxPrice && product.Price >= minPrice);
 
-        var pages = Math.Ceiling((double)productsQuery.Count() / quantity);
 
         if (!string.IsNullOrEmpty(category))
         {
@@ -62,6 +71,9 @@ public class ProductService : IProductService
 
             }
         }
+        
+        var pages = quantity > 0 ? Math.Ceiling((double)productsQuery.Count() / quantity) : 0;
+
 
         // Ограничение по количеству
         var products = await productsQuery
@@ -94,20 +106,20 @@ public class ProductService : IProductService
     {
         return await _context.Products
             .Include(product => product.Reviews)
+            .ThenInclude(review => review.User)
             .FirstOrDefaultAsync(product => product.Id == productId);
     }
 
-    public async Task<Product?> CreateAsync(string name, string description, string category, double price, int quantity, Guid userId)
+    public async Task<Product?> CreateAsync(string name, IFormFile file, string description, string category, double price, int quantity, Guid userId)
     {
-        
-
         if (!await _userService.IsUserAdminAsync(userId)) 
             return null;
 
-
+        
         var newProduct = new Product
         {
 
+            Id = Guid.NewGuid(),
             Name = name,
             Description = description,
             Category = category,
@@ -119,12 +131,17 @@ public class ProductService : IProductService
             UpdatedAt = DateTime.UtcNow
             
         };
+
+
+        newProduct.ImageUrl = await UploadImageAsync(file, newProduct.Id);
+
+
         await _context.Products.AddAsync(newProduct);
         
         return await _context.SaveChangesAsync() > 0 ? newProduct : null;
     }
-
-    public async Task<Product?> UpdateAsync(Guid productId, string name, string description, string category, double price, int quantity,
+    
+    public async Task<Product?> UpdateAsync(Guid productId, string name, IFormFile? file, string description, string category, double price, int quantity,
         Guid userId)
     {
         var product = await GetByIdAsync(productId);
@@ -138,8 +155,17 @@ public class ProductService : IProductService
         product.Quantity = quantity;
 
         product.UpdatedAt = DateTime.UtcNow;
-        
-        
+
+        if (file is not null)
+        {
+            if (!await DeleteImageAsync(product.ImageUrl))
+            {
+                return null;
+            }
+
+            product.ImageUrl = await UploadImageAsync(file, product.Id);
+        }
+
         return await _context.SaveChangesAsync() > 0 ? product : null;
     }
 
@@ -153,8 +179,14 @@ public class ProductService : IProductService
         if (product is null)
             return false;
 
+        if (!await DeleteImageAsync(product.ImageUrl))
+        {
+            return false;
+        }
+
         _context.Products.Remove(product);
         return await _context.SaveChangesAsync() > 0;
+
     }
 
     public async Task<bool> RemoveSomeFromStockAsync(Guid productId, int quantity)
@@ -191,6 +223,13 @@ public class ProductService : IProductService
 
     public async Task<bool> ReCountRatingAsync(Product product)
     {
+        Console.WriteLine("recounting "+ product.Reviews.Count);
+        if (product.Reviews.Count == 0)
+        {
+            product.Rating = 0; // если нет отзывов, рейтинг равен 0 или как-то иначе
+            return await _context.SaveChangesAsync() > 0;
+        }
+        
         double sum = 0;
 
         foreach (var r in product.Reviews)
@@ -198,12 +237,50 @@ public class ProductService : IProductService
             sum += r.Rating;
         }
         product.Rating = sum / product.Reviews.Count;
+        Console.WriteLine(sum / product.Reviews.Count);
+        Console.WriteLine(product.Rating);
+        
+        
+        return await _context.SaveChangesAsync() > 0;
+    }
 
-        if (await _context.SaveChangesAsync() > 0)
+    private async Task<string> UploadImageAsync(IFormFile file, Guid productId)
+    {
+        var containerClient = _blobServiceClient.GetBlobContainerClient(_blobContainerName);
+        await containerClient.CreateIfNotExistsAsync(PublicAccessType.Blob);
+
+        string fileName = $"{productId}_{Guid.NewGuid()}{Path.GetExtension(file.FileName)}";
+        var blobClient = containerClient.GetBlobClient(fileName);
+
+        using (var stream = file.OpenReadStream())
         {
-            return true;
+            await blobClient.UploadAsync(stream, new BlobHttpHeaders { ContentType = file.ContentType });
         }
 
-        return false;
+        return blobClient.Uri.ToString();
+    }
+
+    private async Task<bool> DeleteImageAsync(string imageUrl)
+    {
+        try
+        {
+            var uri = new Uri(imageUrl);
+            var fileName = Path.GetFileName(uri.LocalPath);
+
+            // Получаем клиент для контейнера Blob
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_blobContainerName);
+            var blobClient = containerClient.GetBlobClient(fileName);
+
+            // Удаляем изображение
+            await blobClient.DeleteIfExistsAsync();
+
+            return true;
+        }
+        catch (Exception ex)
+        {
+            // Логируем ошибку
+            Console.WriteLine($"Error deleting blob: {ex.Message}");
+            return false; // Ошибка при удалении
+        }
     }
 }
